[NU-1979] LocalNussknackerWithSingleModel removal
